Complete: Category Management

+++ SEP490_G30_SU23/VBookHaven_Admin/Areas/Admin/Controllers/CategoriesController.cs
@@ -90,7 +90,7 @@ namespace VBookHaven_Admin.Areas.Admin.Controllers
         // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Edit(int id, [Bind("CategoryId,CategoryName,Status")] Category category)
+        public async Task<IActionResult> Edit(int id, [Bind("CategoryId,CategoryName")] Category category)
         {
             if (id != category.CategoryId)
             {
@@ -148,12 +148,17 @@ namespace VBookHaven_Admin.Areas.Admin.Controllers
             {
                 return Problem("Entity set 'VBookHavenDBContext.Categories'  is null.");
             }
-            var category = await _context.Categories.FindAsync(id);
+            var category = await _context.Categories.Include(cate => cate.SubCategories).ThenInclude(sub => sub.Products).SingleOrDefaultAsync( c => c.CategoryId == id);
             if (category != null)
             {
+                if (category.SubCategories.SelectMany(sub => sub.Products).Count() > 0)
+                {
+                    TempData["error"] = "Không thể xóa dữ liệu đang được sử dụng";
+                    return RedirectToAction("Index", "Categories");
+                }
                 _context.Categories.Remove(category);
             }
-
+            TempData["success"] = "Xóa thành công";
             await _context.SaveChangesAsync();
             return RedirectToAction(nameof(Index));
         }
